docs: update docs/index.md, connect.py, execute.py and tutorial.md
